changed int for uint8_t

+++ include/polhemus_ros_driver/viper.hpp
@@ -36,7 +36,7 @@ class Viper: public Polhemus
 {
 public:
   Viper(std::string name, uint16_t rx_buffer_size, uint16_t tx_buffer_size,
-        int sensors_right_glove, int sensors_left_glove);
+        uint8_t sensors_right_glove, uint8_t sensors_left_glove);
   ~Viper(void);
   void device_init();
   int request_num_of_stations(void);
